+++ kurbo/src/shape.rs
@@ -68,12 +68,6 @@ pub trait Shape {
         self.path_elements(tolerance).collect()
     }
 
-    #[deprecated(since = "0.7.0", note = "Use path_elements instead")]
-    #[doc(hidden)]
-    fn to_bez_path(&self, tolerance: f64) -> Self::PathElementsIter<'_> {
-        self.path_elements(tolerance)
-    }
-
     /// Convert into a Bézier path.
     ///
     /// This allocates in the general case, but is zero-cost if the
@@ -87,15 +81,6 @@ pub trait Shape {
         Self: Sized,
     {
         self.to_path(tolerance)
-    }
-
-    #[deprecated(since = "0.7.0", note = "Use into_path instead")]
-    #[doc(hidden)]
-    fn into_bez_path(self, tolerance: f64) -> BezPath
-    where
-        Self: Sized,
-    {
-        self.into_path(tolerance)
     }
 
     /// Returns an iterator over this shape expressed as Bézier path