+++ kurbo/src/translate_scale.rs
@@ -67,13 +67,6 @@ impl TranslateScale {
         TranslateScale::new(translation.into(), 1.0)
     }
 
-    /// Decompose transformation into translation and scale.
-    #[deprecated(note = "use the struct fields directly")]
-    #[inline(always)]
-    pub const fn as_tuple(self) -> (Vec2, f64) {
-        (self.translation, self.scale)
-    }
-
     /// Create a transform that scales about a point other than the origin.
     ///
     /// # Examples